+++ controllers/users.js
@@ -1,6 +1,6 @@
 const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
-const User = require('../models/user');
+const Users = require('../models/user');
 const ConflictError = require('../errors/ConflictError');
 const NotFoundError = require('../errors/NotFoundError');
 const BadRequestError = require('../errors/BadRequestError');
@@ -11,7 +11,7 @@ const createUser = (req, res, next) => {
   const { email, password, name } = req.body;
 
   bcrypt.hash(password, 10)
-    .then((hash) => User.create({
+    .then((hash) => Users.create({
       email, password: hash, name,
     }))
     .then((user) => res.status(201).send({
@@ -29,7 +29,7 @@ const createUser = (req, res, next) => {
 };
 
 const getUser = (req, res, next) => {
-  User.findById(req.user._id)
+  Users.findById(req.user._id)
     .then((user) => {
       if (!user) {
         throw new NotFoundError('Пользователь не найден');
@@ -52,7 +52,7 @@ const getUser = (req, res, next) => {
 const login = (req, res, next) => {
   const { email, password } = req.body;
 
-  User.findUserByCredentials(email, password)
+  Users.findUserByCredentials(email, password)
     .then((user) => {
       const token = jwt.sign(
         { _id: user._id },
